chore: 각 기능 단위로 테스트 코드 나누기
