+++ e2e/playwright.config.ts
@@ -5,7 +5,7 @@ export default defineConfig({
 	fullyParallel: true,
 	forbidOnly: !!process.env.CI,
 	retries: process.env.CI ? 2 : 0,
-	workers: process.env.CI ? 1 : undefined,
+	workers: 1,
 	reporter: [["html", { open: "on-failure" }]],
 	webServer: {
 		command: "pnpm run -w dev:web:preview",
