+++ e2e/tests/fixtures.ts
@@ -1,5 +1,6 @@
 import type { BrowserContext, Page } from "@playwright/test";
 import { test as base, chromium } from "@playwright/test";
+import { PATHS } from "@web-memo/shared/constants";
 import path from "node:path";
 
 process.env.PW_CHROMIUM_ATTACH_TO_OTHER = "1";
@@ -9,6 +10,7 @@ const pathToExtension = path.join(path.resolve(), "..", "dist");
 type ExtensionFixture = {
 	context: BrowserContext;
 	sidePanelPage: Page;
+	loginedPage: Page;
 };
 
 export const test = base.extend<ExtensionFixture>({
@@ -33,8 +35,13 @@ export const test = base.extend<ExtensionFixture>({
 					page.url() ===
 					"chrome-extension://eaiojpmgklfngpjddhoalgcpkepgkclh/side-panel/index.html",
 			)!;
-
 		use(sidePanelPage);
+	},
+	loginedPage: async ({ page }, use) => {
+		await page.goto("/en/login");
+		await page.getByTestId("test-login-button").click();
+		await page.waitForURL(new RegExp(PATHS.memos));
+		use(page);
 	},
 	baseURL: "http://localhost:3000",
 });
