+++ e2e/tests/sidePanel.test.ts
