+++ packages/shared/src/utils/Environment.ts
@@ -1,6 +1,5 @@
 import { CONFIG } from "@web-memo/env";
 
-export const isServer = () => typeof window === "undefined";
 export const isMac = () =>
 	typeof navigator !== "undefined" &&
 	/Mac|iPhone|iPad|iPod/.test(navigator.platform);
