+++ packages/web/src/app/[lng]/(no-auth)/login/_components/LoginSection/index.tsx
@@ -25,6 +25,7 @@ export default async function LoginSection({ lng }: LoginSectionProps) {
 				<Button
 					formAction={signInWithOAuth.bind(null, "kakao")}
 					className="h-12 bg-[rgb(247,228,76)] text-black hover:bg-[rgb(247,228,76)]"
+					data-testid="kakao-login-button"
 				>
 					<Image
 						src="/images/svgs/kakao.svg"
@@ -37,6 +38,7 @@ export default async function LoginSection({ lng }: LoginSectionProps) {
 				<Button
 					formAction={signInWithOAuth.bind(null, "google")}
 					className="h-12 bg-white text-black hover:bg-white"
+					data-testid="google-login-button"
 				>
 					<Image
 						src="/images/svgs/google.svg"
@@ -54,6 +56,7 @@ export default async function LoginSection({ lng }: LoginSectionProps) {
 							SUPABASE.testPassword,
 						)}
 						className="h-12 bg-green-300 text-black hover:bg-green-300"
+						data-testid="test-login-button"
 					>
 						{t("login.testLogin")}
 					</Button>
